+++ src/prefPages/toggles.ts
@@ -12,32 +12,38 @@ import {
 	UpDownButton,
 	Row,
 	Group,
+	DialogRow,
+	ToggleButtonRow,
+	ResetButton,
 	setScrollToFocus,
 	delayedSetScrollToFocus,
 	fixPageScrollIssue,
 } from "../libs/prefComponents.js"
 
-function ToggleEditorGroup(settings: Gio.Settings, page: Adw.PreferencesPage): Adw.PreferencesGroup {
-	const systemToggleNames = ToggleEditorGroup.getSystemToggleNames()
+// #region ToggleOrderGroup
+function ToggleOrderGroup(settings: Gio.Settings, page: Adw.PreferencesPage): Adw.PreferencesGroup {
+	const systemToggleNames = ToggleOrderGroup.getSystemToggleNames()
 	const itemRows = new Map<QuickToggleOrderItem, Adw.ActionRow>()
 
 	const group = Group({
+		title: _("Ordering and Hiding"),
+		headerSuffix: ResetButton({ settings, bind: "toggle-order", marginBottom: 0, marginTop: 0 }),
 	})
 
 	const setHide = (item: QuickToggleOrderItem, hide: boolean)=>{
-		const list = ToggleEditorGroup.getOrderListFromSettings(settings)
+		const list = ToggleOrderGroup.getOrderListFromSettings(settings)
 		list.find(targetItem => QuickToggleOrderItem.match(targetItem, item)).hide = hide
-		ToggleEditorGroup.setOrderListToSettings(settings, list)
+		ToggleOrderGroup.setOrderListToSettings(settings, list)
 	}
 	const move = (item: QuickToggleOrderItem, direction: UpDownButton.Direction)=>{
-		const list = ToggleEditorGroup.getOrderListFromSettings(settings)
+		const list = ToggleOrderGroup.getOrderListFromSettings(settings)
 		const index = list.findIndex(targetItem => QuickToggleOrderItem.match(targetItem, item))
 		const targetIndex = index + (direction == UpDownButton.Direction.Up ? -1 : 1)
 		if (targetIndex < 0 || targetIndex >= list.length) return
 		const row = list[index]
 		list[index] = list[targetIndex]
 		list[targetIndex] = row
-		ToggleEditorGroup.setOrderListToSettings(settings, list)
+		ToggleOrderGroup.setOrderListToSettings(settings, list)
 	}
 	const removeOrphans = (list: QuickToggleOrderItem[])=>{
 		for (const [targetItem, row] of itemRows.entries()) {
@@ -52,8 +58,8 @@ function ToggleEditorGroup(settings: Gio.Settings, page: Adw.PreferencesPage): A
 				continue
 			const row = Row({
 				settings,
-				title: ToggleEditorGroup.getDisplayName(newItem, systemToggleNames),
-				subtitle: ToggleEditorGroup.getSubtitle(newItem),
+				title: ToggleOrderGroup.getDisplayName(newItem, systemToggleNames),
+				subtitle: ToggleOrderGroup.getSubtitle(newItem),
 				sensitiveBind: "toggle-order-enabled"
 			})
 
@@ -93,7 +99,7 @@ function ToggleEditorGroup(settings: Gio.Settings, page: Adw.PreferencesPage): A
 
 	const update = ()=>{
 		setScrollToFocus(page, false)
-		const list = ToggleEditorGroup.getOrderListFromSettings(settings)
+		const list = ToggleOrderGroup.getOrderListFromSettings(settings)
 		pushChildren(list)
 		removeOrphans(list)
 		orderChildren(list)
@@ -105,7 +111,7 @@ function ToggleEditorGroup(settings: Gio.Settings, page: Adw.PreferencesPage): A
 
 	return group
 }
-namespace ToggleEditorGroup {
+namespace ToggleOrderGroup {
 	export function getSystemToggleNames(): Map<string, string> {
 		const IGNORE_XGETTEXT=_
 		return new Map<string, string>([
@@ -165,11 +171,140 @@ namespace ToggleEditorGroup {
 		return ""
 	}
 }
+// #endregion ToggleOrderGroup
+
+// #region SystemItemOrderGroup
+function SystemItemOrderGroup(settings: Gio.Settings, page: Adw.PreferencesPage): Adw.PreferencesGroup {
+	let items = new Map<string, Adw.ActionRow>()
+	let group: Adw.PreferencesGroup
+	const reorder = ()=>{
+		setScrollToFocus(page, false)
+		const order = SystemItemOrderGroup.copyOrder(settings.get_strv("system-items-order"))
+		for (const name of order) {
+			const target = items.get(name)
+			group.remove(target)
+			group.add(target)
+		}
+		delayedSetScrollToFocus(page, true)
+	}
+	const move = (direction: UpDownButton.Direction, name: string)=>{
+		const order = SystemItemOrderGroup.copyOrder(settings.get_strv("system-items-order"))
+		const index = order.indexOf(name)
+		if (direction == UpDownButton.Direction.Up) {
+			if (index == 0) return
+			order[index] = order[index - 1]
+			order[index - 1] = name
+		} else {
+			if (index == (SystemItemOrderGroup.DefaultOrder.length - 1)) return
+			order[index] = order[index + 1]
+			order[index + 1] = name
+		}
+		settings.set_strv("system-items-order", order)
+	}
+
+	const orderConnection = settings.connect("changed::system-items-order", reorder)
+	page.connect("destroy", ()=>settings.disconnect(orderConnection))
+
+	return Group({
+		title: _("Ordering and Hiding"),
+		headerSuffix: ResetButton({ settings, bind: "system-items-order", marginBottom: 0, marginTop: 0 }),
+		onCreated(row: Adw.PreferencesGroup) {
+			group = row
+			reorder()
+		},
+	}, [
+		Row({
+			title: _("Desktop Spacer"),
+			prefix: UpDownButton({
+				settings,
+				sensitiveBind: "system-items-enabled",
+				action: (direction) => move(direction, "desktopSpacer"),
+			}),
+			onCreated(row) {
+				items.set("desktopSpacer", row)
+			},
+		}),
+		Row({
+			title: _("Laptop Spacer"),
+			prefix: UpDownButton({
+				settings,
+				sensitiveBind: "system-items-enabled",
+				action: (direction) => move(direction, "laptopSpacer"),
+			}),
+			onCreated(row) {
+				items.set("laptopSpacer", row)
+			},
+		}),
+		...[
+			{
+				title: _("Capture button"),
+				bind: "system-items-hide-screenshot",
+				icon: "camera-photo",
+				targetName: "screenshot",
+			},
+			{
+				title: _("Settings button"),
+				bind: "system-items-hide-settings",
+				icon: "preferences-system-symbolic",
+				targetName: "settings",
+			},
+			{
+				title: _("Lock button"),
+				bind: "system-items-hide-lock",
+				icon: "system-lock-screen-symbolic",
+				targetName: "lock",
+			},
+			{
+				title: _("Shutdown button"),
+				bind: "system-items-hide-shutdown",
+				icon: "system-shutdown-symbolic",
+				targetName: "shutdown",
+			},
+			{
+				title: _("Battery button"),
+				bind: "system-items-hide-battery",
+				icon: "battery-symbolic",
+				targetName: "battery",
+			},
+		].map(item => ToggleButtonRow({
+			settings,
+			text: _("Hide"),
+			sensitiveBind: "system-items-enabled",
+			...item,
+			onCreated(row: Adw.ActionRow) {
+				items.set(item.targetName, row)
+				row.add_prefix(new Gtk.Image({
+					icon_name: item.icon,
+					pixel_size: 16,
+					margin_start: 8
+				}))
+				row.add_prefix(UpDownButton({
+					settings,
+					sensitiveBind: "system-items-enabled",
+					action: (direction) => move(direction, item.targetName),
+				}))
+			},
+		}))
+	])
+}
+namespace SystemItemOrderGroup {
+	export const DefaultOrder = ['battery', 'laptopSpacer', 'screenshot', 'settings', 'desktopSpacer', 'lock', 'shutdown']
+	export function copyOrder(order: string[]): string[] {
+		return DefaultOrder
+		.map(item => ({
+			item,
+			index: order.indexOf(item),
+		}))
+		.sort((a, b)=>a.index-b.index)
+		.map(item => item.item)
+	}
+}
+// #endregion SystemItemOrderGroup
 
 export const TogglesPage = GObject.registerClass({
 	GTypeName: Config.baseGTypeName+'TogglesPage',
 }, class TogglesPage extends Adw.PreferencesPage {
-	constructor(settings: Gio.Settings, _prefs: QstExtensionPreferences, _window: Adw.PreferencesWindow) {
+	constructor(settings: Gio.Settings, _prefs: QstExtensionPreferences, window: Adw.PreferencesWindow) {
 		super({
 			name: 'Toggles',
 			title: _('Toggles'),
@@ -180,8 +315,8 @@ export const TogglesPage = GObject.registerClass({
 		// Add
 		Group({
 			parent: this,
-			title: _('Add more buttons'),
-			description: _('Turn on the buttons you want to add on Quick Settings'),
+			title: _('Add more toggles'),
+			description: _('Turn on the toggles you want to add on Quick Settings'),
 		}, [
 			SwitchRow({
 				settings,
@@ -197,17 +332,55 @@ export const TogglesPage = GObject.registerClass({
 			}),
 		])
 
-		// Order
+		// Quick toggles
 		Group({
 			parent: this,
-			title: _("Ordering and Hiding"),
-			description: _("Reorder and hide quick toggles"),
+			title: _("Quick Toggles Layout"),
+			description: _("Adjust quick toggles layout"),
 			headerSuffix: SwitchRow({
 				settings,
 				bind: "toggle-order-enabled",
 			}),
 		},[
-			ToggleEditorGroup( settings, this )
+			DialogRow({
+				settings,
+				window,
+				sensitiveBind: "toggle-order-enabled",
+				title: _("Ordering and Hiding"),
+				subtitle: _("Reorder and hide quick toggles"),
+				dialogTitle: _("Adjust system items layout"),
+				experimental: true,
+				childrenRequest: page => [ToggleOrderGroup(settings, page)],
+			}),
+		])
+
+		// System Items
+		Group({
+			parent: this,
+			title: _("System Items Layout"),
+			headerSuffix: SwitchRow({
+				settings,
+				bind: "system-items-enabled",
+			}),
+			description: _("Adjust system items layout"),
+		},[
+			SwitchRow({
+				settings,
+				title: _("Hide layout"),
+				subtitle: _("Hide all buttons and layout box"),
+				bind: "system-items-hide",
+				sensitiveBind: "system-items-enabled",
+			}),
+			DialogRow({
+				settings,
+				window,
+				sensitiveBind: "system-items-enabled",
+				title: _("Ordering and Hiding"),
+				subtitle: _("Hide all buttons and layout box"),
+				dialogTitle: _("Adjust system items layout"),
+				experimental: true,
+				childrenRequest: page => [SystemItemOrderGroup(settings, page)],
+			}),
 		])
 	}
 })